requirements.txt
flask
pytest
mysql-connector-python
python-dotenv
